requirements.txt
pyyaml
sqlachemy
pandas
Office365-REST-Python-Client
